MAINT: Moved functions to array_operations and modified accum
