+++ src/porepy/examples/flow_benchmark_2d_case_1.py
@@ -1,3 +1,4 @@
+
 """This module contains the implementation of Case 1 from the 2D flow benchmark [1].
 
 We provide the two variants of this benchmark, i.e., Case 1a (conductiv fractures) and
