+++ src/porepy/fracs/meshing.py
@@ -19,7 +19,6 @@ from porepy.fracs import split_grid, structured
 from porepy.grids import mortar_grid
 from porepy.grids.md_grid import MixedDimensionalGrid
 from porepy.numerics.linalg.matrix_operations import sparse_array_to_row_col_data
-from porepy.utils import mcolon
 
 logger = logging.getLogger(__name__)
 
@@ -282,7 +281,7 @@ def _tag_faces(grids, check_highest_dim=True):
                 # We find the global nodes of all boundary faces
                 bnd_faces_l = g.get_all_boundary_faces()
                 indptr = g.face_nodes.indptr
-                fn_loc = mcolon.mcolon(indptr[bnd_faces_l], indptr[bnd_faces_l + 1])
+                fn_loc = pp.array_operations.mcolon(indptr[bnd_faces_l], indptr[bnd_faces_l + 1])
                 nodes_loc = g.face_nodes.indices[fn_loc]
                 # Convert to global numbering
                 nodes_glb = g.global_point_ind[nodes_loc]
