+++ src/porepy/fracs/wells_3d.py
@@ -615,7 +615,7 @@ def compute_well_rock_matrix_intersections(
         second = sd_w_cn.indptr[sd_w_cells + 1]
 
         nodes_w.append(
-            sd_w_cn.indices[pp.utils.mcolon.mcolon(first, second)].reshape((-1, 2)).T
+            sd_w_cn.indices[pp.array_operations.mcolon(first, second)].reshape((-1, 2)).T
         )
 
     # Operate on the rock matrix grid.
