+++ src/porepy/grids/grid.py
@@ -25,7 +25,7 @@ from scipy import sparse as sps
 
 import porepy as pp
 from porepy.numerics.linalg.matrix_operations import sparse_array_to_row_col_data
-from porepy.utils import mcolon, tags
+from porepy.utils import tags
 
 
 class Grid:
@@ -866,7 +866,7 @@ class Grid:
             if faces.size > 0:
                 first = self.face_nodes.indptr[faces]
                 second = self.face_nodes.indptr[faces + 1]
-                nodes = self.face_nodes.indices[mcolon.mcolon(first, second)]
+                nodes = self.face_nodes.indices[pp.array_operations.mcolon(first, second)]
                 self.tags[node_tag][nodes] = True
 
     def cell_diameters(self, cn: Optional[sps.spmatrix] = None) -> np.ndarray:
@@ -1119,8 +1119,8 @@ class Grid:
         # Expand the indices to the correct size. The trace operator is a mapping from
         # cells to faces, so we need boundary cells in the columns and boundary faces in
         # the rows.
-        rows = pp.fvutils.expand_indices_nd(bound_faces, dim)
-        cols = pp.fvutils.expand_indices_nd(bound_cells, dim)
+        rows = pp.array_operations.expand_indices_nd(bound_faces, dim)
+        cols = pp.array_operations.expand_indices_nd(bound_cells, dim)
         trace = sps.coo_matrix(
             (np.ones(bound_faces.size * dim), (rows, cols)),
             shape=(self.num_faces * dim, self.num_cells * dim),
