+++ src/porepy/numerics/fv/mpsa.py
@@ -379,7 +379,7 @@ class Mpsa(Discretization):
         )
 
         if update:
-            update_ind = pp.fvutils.expand_indices_nd(active_faces, sd.dim)
+            update_ind = pp.array_operations.expand_indices_nd(active_faces, sd.dim)
             matrix_dictionary[self.stress_matrix_key][update_ind] = stress_glob[
                 update_ind
             ]
@@ -1636,7 +1636,7 @@ class Mpsa(Discretization):
 
             # Distribute (relevant parts of) Hook's law on subcells This will be nd
             # rows, thus cell ci is associated with indices ci*nd+np.arange(nd)
-            sub_cell_ind = pp.fvutils.expand_indices_nd(cell_node_blocks[0], nd)
+            sub_cell_ind = pp.array_operations.expand_indices_nd(cell_node_blocks[0], nd)
             sym_vals = sym_dim[sub_cell_ind]
             asym_vals = asym_dim[sub_cell_ind]
