+++ tests/numerics/fv/test_mpsa.py
@@ -137,7 +137,7 @@ def test_partial_discretization_specified_nodes(cell_id: int, discretization_mat
     ]
 
     active_faces = data[pp.PARAMETERS][keyword]["active_faces"]
-    active_faces_nd = pp.fvutils.expand_indices_nd(active_faces, g.dim)
+    active_faces_nd = pp.array_operations.expand_indices_nd(active_faces, g.dim)
 
     for partial, full in zip(
         [partial_stress, partial_bound],
@@ -194,7 +194,7 @@ def test_partial_discretization_one_cell_at_a_time():
         active_faces = data[pp.PARAMETERS][keyword]["active_faces"]
 
         if np.any(faces_covered):
-            del_faces = pp.fvutils.expand_indices_nd(np.where(faces_covered)[0], g.dim)
+            del_faces = pp.array_operations.expand_indices_nd(np.where(faces_covered)[0], g.dim)
             pp.fvutils.remove_nonlocal_contribution(
                 del_faces, 1, partial_stress, partial_bound
             )
